Create cc-6.c
**Problem**
Each item in Chef’s menu is assigned a spice level from 11 to 1010. Based on the spice level, the item is categorised as:
MILD: If the spice level is less than 44.
MEDIUM: If the spice level is greater than equal to 44 but less than 77.
HOT: If the spice level is greater than equal to 77.