Create cc-9.c
Problem
After the phenomenal success of the 36th Chamber of Shaolin, 
San Te has decided to start 37th Chamber of Shaolin. 
The aim this time is to equip women with shaolin self-defence techniques.
The only condition for a woman to be eligible for the special training is that she must be between
 1010 and 6060 years of age, inclusive of both 10and 60.
Given the ages of NN women in his village, please help San Te 
find out how many of them are eligible for the special training.
Input Format
The first line of input contains a single integer T,
 denoting the number of test cases.
 The description of T test cases follows.
The first line of each test case contains a single integer N,
 the number of women.
The second line of each test case contains N space-separated integers A_1, A_2,..., NA 
the ages of the women.